requirements.txt
torch==1.10.0
torchvision==0.11.1
torchaudio==0.10.0
# numpy==1.21.2
gym==0.22.0
pdf2image==1.16.0
Pillow==8.3.2
pytesseract
openai
pandas
python-dotenv
ipython
langchain
faiss-cpu
termcolor
openpyxl
unstructured
networkx
tiktoken
scikit-learn
ultralytics
streamlit 
pdf2image
shutil
google-cloud-vision #pip3でインストールしないとダメかも
# Ultralytics requirements
# Example: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.3.0
numpy==1.22.2 # pinned by Snyk to avoid a vulnerability
opencv-python>=4.6.0
pillow>=7.1.2
pyyaml>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.8.0
torchvision>=0.9.0
tqdm>=4.64.0
# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0
